change config and add val() test() for trainer
