--- cfg/layer_compressible_mod567asp.json
@@ -1,240 +1,0 @@
-{
-    "name": "KnowledgeDistillation",
-    "n_gpu": 1,
-    "teacher": {
-        "type": "DeepWV3Plus",
-        "args": {
-            "num_classes": 19
-        },
-        "snapshot": "checkpoints/cityscapes_best.pth"
-    },
-    "transforms": {
-        "joint_transforms": {
-            "crop_size": 512,
-            "scale_min": 0.5,
-            "scale_max": 2,
-            "ignore_label": 255
-        },
-        "extended_transforms": {
-            "color_aug": 0.2,
-            "blur": "gaussian"
-        }
-    },
-    "train_data_loader": {
-        "type": "CityscapesDataloader",
-        "args": {
-            "data_dir": "data/",
-            "batch_size": 4,
-            "shuffle": true,
-            "validation_split": 0.0,
-            "num_workers": 0,
-            "split": "train",
-            "mode": "fine",
-            "target_type": "semantic",
-            "num_samples": 528
-        }
-    },
-    "val_data_loader": {
-        "type": "CityscapesDataloader",
-        "args": {
-            "data_dir": "data/",
-            "batch_size": 1,
-            "shuffle": false,
-            "validation_split": 0,
-            "num_workers": 0,
-            "split": "val",
-            "mode": "fine",
-            "target_type": "semantic",
-            "num_samples": 50
-        }
-    },
-    "optimizer": {
-        "type": "RAdam",
-        "args": {
-            "lr": 0.005
-        }
-    },
-    "supervised_loss": {
-        "type": "FocalLoss",
-        "args": {
-            "gamma": 2,
-            "reduction": "mean",
-            "ignore_index": 255
-        }
-    },
-    "kd_loss": {
-        "type": "MSELoss",
-        "args": {
-            "reduction": "mean",
-            "num_classes": 1
-        }
-    },
-    "hint_loss": {
-        "type": "MSELoss",
-        "args": {
-            "reduction": "mean",
-            "num_classes": 1
-        }
-    },
-    "metrics": [],
-    "lr_scheduler": {
-        "type": "MyReduceLROnPlateau",
-        "args": {
-            "mode": "min",
-            "threshold": 0.1,
-            "factor": 0.5,
-            "patience": 0,
-            "verbose": true,
-            "min_lr": 0.00001,
-            "threshold_mode": "rel"
-        }
-    },
-    "trainer": {
-        "name": "LayerCompressibleTrainer",
-        "epochs": 2,
-        "save_dir": "saved/",
-        "save_period": 5,
-        "verbosity": 2,
-        "monitor": "min val_mIoU",
-        "early_stop": 1000,
-        "accumulation_steps": 1,
-        "log_step": 5,
-        "do_validation_interval": 5,
-        "lr_scheduler_step_interval": 15,
-        "len_epoch": 50,
-        "tensorboard": true
-    },
-    "pruning": {
-        "pruner": {
-            "dilation": 2,
-            "padding": 2,
-            "kernel_size": 3
-        },
-        "compress_rate": 0.2,
-        "auxiliary_num_layers": 6,
-        "pruning_plan": [
-            {
-                "name": "mod5.block1.convs.conv1",
-                "compress_rate": 0.3,
-                "epoch": 40
-            },
-            {
-                "name": "mod5.block2.convs.conv1",
-                "compress_rate": 0.2,
-                "epoch": 40
-            },
-            {
-                "name": "mod5.block2.convs.conv2",
-                "compress_rate": 0.2,
-                "epoch": 40
-            },
-            {
-                "name": "mod5.block3.convs.conv1",
-                "compress_rate": 0.2,
-                "epoch": 40
-            },
-            {
-                "name": "mod5.block3.convs.conv2",
-                "compress_rate": 0.2,
-                "epoch": 70
-            },
-            {
-                "name": "mod6.block1.convs.conv2",
-                "compress_rate": 0.2,
-                "epoch": 70
-            },
-            {
-                "name": "mod7.block1.convs.conv1",
-                "compress_rate": 0.2,
-                "epoch": 70
-            },
-            {
-                "name": "mod7.block1.convs.conv2",
-                "compress_rate": 0.15,
-                "epoch": 20
-            },
-            {
-                "name": "mod7.block1.convs.conv3",
-                "compress_rate": 0.1,
-                "epoch": 20
-            },
-            {
-                "name": "aspp.features.1.0",
-                "compress_rate": 0.2,
-                "epoch": 1
-            },
-            {
-                "name": "aspp.features.2.0",
-                "compress_rate": 0.2,
-                "epoch": 1
-            },
-            {
-                "name": "aspp.features.3.0",
-                "compress_rate": 0.2,
-                "epoch": 1
-            }
-        ]
-    },
-    "teaching_assistant": {
-        "interval": 15,
-        "tol": 0.005
-    },
-    "weight_scheduler": {
-        "alpha": {
-            "value": 0.0001,
-            "anneal_rate": 2,
-            "max": 0
-        },
-        "beta": {
-            "value": 0.999,
-            "anneal_rate": 0.9,
-            "min": 0.999
-        },
-        "gamma": {
-            "value": 1,
-            "anneal_rate": 1
-        }
-    },
-    "layer_compressible": [
-      {
-        "layer_name": "aspp.features.3.0",
-        "compress_rate": 0,
-        "lrs": [1, 0.5, 0.1, 0.05, 0.01, 0.005, 0.001, 0.0005, 0.0001]
-      },
-      {
-        "layer_name": "aspp.features.2.0",
-        "compress_rate": 0,
-        "lrs": [1, 0.5, 0.1, 0.05, 0.01, 0.005, 0.001, 0.0005, 0.0001]
-      },
-      {
-        "layer_name": "aspp.features.1.0",
-        "compress_rate": 0,
-        "lrs": [1, 0.5, 0.1, 0.05, 0.01, 0.005, 0.001, 0.0005, 0.0001]
-      },
-      {
-        "layer_name": "mod6.block1.convs.conv2",
-        "compress_rate": 0,
-        "lrs": [1, 0.5, 0.1, 0.05, 0.01, 0.005, 0.001, 0.0005, 0.0001]
-      },
-      {
-        "layer_name": "mod5.block3.convs.conv2",
-        "compress_rate": 0,
-        "lrs": [1, 0.5, 0.1, 0.05, 0.01, 0.005, 0.001, 0.0005, 0.0001]
-      },
-      {
-        "layer_name": "mod5.block3.convs.conv1",
-        "compress_rate": 0,
-        "lrs": [1, 0.5, 0.1, 0.05, 0.01, 0.005, 0.001, 0.0005, 0.0001]
-      },
-      {
-        "layer_name": "mod5.block2.convs.conv2",
-        "compress_rate": 0,
-        "lrs": [1, 0.5, 0.1, 0.05, 0.01, 0.005, 0.001, 0.0005, 0.0001]
-      },
-      {
-        "layer_name": "mod5.block1.convs.conv1",
-        "compress_rate": 0,
-        "lrs": [1, 0.5, 0.1, 0.05, 0.01, 0.005, 0.001, 0.0005, 0.0001]
-      }
-    ]
-}
