--- cfg/layerwise.json
@@ -1,290 +1,0 @@
-{
-    "name": "KnowledgeDistillation",
-    "n_gpu": 1,
-    "teacher": {
-        "type": "DeepWV3Plus",
-        "args": {
-            "num_classes": 19
-        },
-        "snapshot": "checkpoints/cityscapes_best.pth"
-    },
-    "transforms": {
-        "joint_transforms": {
-            "crop_size": 512,
-            "scale_min": 0.5,
-            "scale_max": 2,
-            "ignore_label": 255
-        },
-        "extended_transforms": {
-            "color_aug": 0.2,
-            "blur": "gaussian"
-        }
-    },
-    "train_data_loader": {
-        "type": "CityscapesDataloader",
-        "args": {
-            "data_dir": "data/",
-            "batch_size": 4,
-            "shuffle": true,
-            "validation_split": 0.0,
-            "num_workers": 0,
-            "split": "train",
-            "mode": "fine",
-            "target_type": "semantic"
-        }
-    },
-    "val_data_loader": {
-        "type": "CityscapesDataloader",
-        "args": {
-            "data_dir": "data/",
-            "batch_size": 1,
-            "shuffle": false,
-            "validation_split": 0,
-            "num_workers": 0,
-            "split": "val",
-            "mode": "fine",
-            "target_type": "semantic",
-            "num_samples": 50
-        }
-    },
-    "test_data_loader": {
-        "type": "CityscapesDataloader",
-        "args": {
-            "data_dir": "data/",
-            "batch_size": 1,
-            "shuffle": false,
-            "validation_split": 0,
-            "num_workers": 0,
-            "split": "test",
-            "mode": "fine",
-            "target_type": "semantic",
-            "num_samples": 50
-        }
-    },
-    "optimizer": {
-        "type": "RAdam",
-        "args": {
-            "lr": 0.005
-        }
-    },
-    "supervised_loss": {
-        "type": "CrossEntropyLoss2d",
-        "args": {
-            "ignore_index": 255
-        }
-    },
-    "kd_loss": {
-        "type": "MSELoss",
-        "args": {
-            "reduction": "mean",
-            "num_classes": 1
-        }
-    },
-    "hint_loss": {
-        "type": "MSELoss",
-        "args": {
-            "reduction": "mean",
-            "num_classes": 1000
-        }
-    },
-    "metrics": [],
-    "lr_scheduler": {
-        "type": "StepLR",
-        "args": {
-            "step_size": 1,
-            "gamma": 0.8
-        }
-    },
-    "trainer": {
-        "name": "LayerwiseTrainer",
-        "epochs": 1000,
-        "save_dir": "saved/",
-        "save_period": 1,
-        "verbosity": 2,
-        "monitor": "min val_mIoU",
-        "early_stop": 1000,
-        "accumulation_steps": 1,
-        "log_step": 5,
-        "do_validation_interval": 100,
-        "lr_scheduler_step_interval": 15,
-        "len_epoch": 100,
-        "tensorboard": true
-    },
-    "pruning": {
-        "pruner": {
-            "dilation": 5,
-            "padding": 20,
-            "kernel_size": 9
-        },
-        "pruning_plan": [
-            {
-                "name": "mod4.block2.convs.conv2",
-                "epoch": 1
-            },
-            {
-                "name": "mod4.block3.convs.conv1",
-                "epoch": 4
-            },
-            {
-                "name": "mod4.block3.convs.conv2",
-                "epoch": 7
-            },
-            {
-                "name": "mod4.block4.convs.conv2",
-                "epoch": 10
-            },
-            {
-                "name": "mod4.block5.convs.conv2",
-                "epoch": 13
-            },
-            {
-                "name": "mod4.block6.convs.conv2",
-                "epoch":  16
-            },
-            {
-                "name": "mod5.block2.convs.conv2",
-                "epoch": 19
-            },
-            {
-                "name": "mod7.block1.convs.conv2",
-                "epoch": 22
-            },
-            {
-                "name": "aspp.features.1.0",
-                "epoch": 25
-            },
-            {
-                "name": "aspp.features.2.0",
-                "epoch": 28
-            },
-            {
-                "name": "aspp.features.3.0",
-                "epoch": 31
-            }
-        ],
-        "hint": [
-            {
-                "name": "mod4.block3",
-                "epoch": 1
-            },
-            {
-                "name": "mod4.block4",
-                "epoch": 4
-            },
-            {
-                "name": "mod4.block4",
-                "epoch": 7
-            },
-            {
-                "name": "mod4.block5",
-                "epoch": 10
-            },
-            {
-                "name": "mod4.block6",
-                "epoch": 13
-            },
-            {
-                "name": "mod5.block1",
-                "epoch": 16
-            },
-            {
-                "name": "mod6.block1",
-                "epoch": 19
-            },
-            {
-                "name": "mod7.block1",
-                "epoch": 22
-            },
-            {
-                "name": "aspp",
-                "epoch": 25
-            },
-            {
-                "name": "aspp",
-                "epoch": 28
-            },
-            {
-                "name": "aspp",
-                "epoch": 31
-            }
-        ],
-        "unfreeze": [
-            {
-                "name": "mod4.block2.convs.conv2",
-                "epoch": 1
-            },
-            {
-                "name": "mod4.block3.convs.conv1",
-                "epoch": 4
-            },
-            {
-                "name": "mod4.block3.convs.conv2",
-                "epoch": 7
-            },
-            {
-                "name": "mod4.block4.convs",
-                "epoch": 10
-            },
-            {
-                "name": "mod4.block5.convs",
-                "epoch": 13
-            },
-            {
-                "name": "mod4.block6.convs",
-                "epoch": 16
-            },
-            {
-                "name": "mod5.block2.convs",
-                "epoch": 19
-            },
-            {
-                "name": "mod7.block1.convs.conv1",
-                "epoch": 22
-            },
-            {
-                "name": "mod7.block1.convs.conv2",
-                "epoch": 22
-            },
-            {
-                "name": "aspp.features.1.0",
-                "epoch": 25
-            },
-            {
-                "name": "aspp.features.2.0",
-                "epoch": 28
-            },
-            {
-                "name": "aspp.features.3.0",
-                "epoch": 31
-            }
-        ]
-    },
-    "weight_scheduler": {
-        "alpha": {
-            "value": 0.0001,
-            "anneal_rate": 2,
-            "max": 0
-        },
-        "beta": {
-            "value": 0.99,
-            "anneal_rate": 0.95,
-            "min": 0.99
-        },
-        "gamma": {
-            "value": 1,
-            "anneal_rate": 1
-        }
-    },
-    "test": {
-        "type": "sliding",
-        "args": {
-            "scales": [1.0],
-            "crop_size": 1024
-        }
-    },
-    "submission": {
-        "save_output": true,
-        "path_output": "submission/",
-        "ext": "png"
-    }
-}
